Remove Element::IfTags / Element::IfCategory, check conditions in parse.
